+++ apecs-physics/src/Apecs/Physics.hs
@@ -12,7 +12,7 @@ module Apecs.Physics (
 
   -- * Body
   Body (..), Position (..), Velocity (..), Angle (..), AngularVelocity (..), Force (..),
-  BodyMass (..), Moment (..), CenterOfGravity (..), Torque (..),
+  BodyMass (..), Moment (..), CenterOfGravity (..), Torque (..), Shapes (..), Constraints (..),
 
   -- * Shape
   Convex (..), Shape (Shape, ShapeExtend),